feat(members): refactor to use mock data and utility functions for role and status
